refactor(mjh/types): split 3 multi-interface type files into one-per-file
Six interfaces co-located in three files violate the project convention
(one type/interface per file in src/types/). Split into 6 separate files:

- discovered-job.ts (DiscoveredJob — no change)
- discovered-job-list-response.ts (new, extracted from discovered-job.ts)
- discovery-source.ts (DiscoverySource — no change)
- discovery-source-create-request.ts (new, extracted from discovery-source.ts)
- profile/profile.ts (Profile — no change, now imports DiscoveryDefaults)
- profile/discovery-defaults.ts (new, extracted from profile.ts)

Updated 3 consumer import paths: store/discoverApi.ts,
types/profile/profile.ts, types/profile/profile-update-request.ts.
Resolves TECH_DEBT.md MEDIUM "One-type-per-file convention violated in 3 files".
