+++ apps/myjobhunter/frontend/src/types/discovery/discovered-job-list-response.ts
@@ -1,0 +1,7 @@
+import type { DiscoveredJob } from "./discovered-job";
+
+export interface DiscoveredJobListResponse {
+  items: DiscoveredJob[];
+  total: number;
+  state: "inbox" | "saved" | "all";
+}
