+++ apps/myjobhunter/frontend/src/types/discovery/discovery-source-create-request.ts
@@ -1,0 +1,6 @@
+/** Body for POST /discover/sources. */
+export interface DiscoverySourceCreate {
+  source: string;
+  config: Record<string, unknown>;
+  fetch_interval_minutes?: number;
+}
