+++ apps/myjobhunter/frontend/src/types/discovery/discovery-source.ts
@@ -13,10 +13,3 @@ export interface DiscoverySource {
   created_at: string;
   updated_at: string;
 }
-
-/** Body for POST /discover/sources. */
-export interface DiscoverySourceCreate {
-  source: string;
-  config: Record<string, unknown>;
-  fetch_interval_minutes?: number;
-}
