+++ apps/myjobhunter/frontend/src/types/profile/profile-update-request.ts
@@ -1,4 +1,4 @@
-import type { DiscoveryDefaults } from "./profile";
+import type { DiscoveryDefaults } from "./discovery-defaults";
 
 /**
  * Body for PATCH /profile. All fields optional.
