+++ apps/myjobhunter/frontend/src/types/profile/profile.ts
@@ -1,3 +1,5 @@
+import type { DiscoveryDefaults } from "./discovery-defaults";
+
 /**
  * Profile as returned by GET /profile and PATCH /profile.
  * Mirrors ProfileResponse in backend/app/schemas/profile/profile_response.py.
@@ -30,17 +32,4 @@ export interface Profile {
 
   created_at: string;
   updated_at: string;
-}
-
-export interface DiscoveryDefaults {
-  excluded_industry_chips?: string[];
-  excluded_keywords?: string[];
-  employment_type?: string;
-  experience?: string;
-  country?: string;
-  date_posted?: string;
-  // Phase C scoring inputs (written by Phase C).
-  preferred_industries?: string[];
-  preferred_stack?: string[];
-  rejected_stack?: string[];
 }
